refactor: move attribute to web component
