+++ src/app/resume-page/web-resume/attribute/attribute.component.ts
