+++ src/app/resume-page/web-resume/education-section/education/education.component.spec.ts
@@ -12,7 +12,7 @@ import { byTestId } from '@/test/helpers/test-id'
 import { CardHeaderComponent } from '../../../card/card-header/card-header.component'
 import { CardHeaderTextsComponent } from '../../../card/card-header/card-header-texts/card-header-texts.component'
 import { CardHeaderAttributesComponent } from '../../../card/card-header/card-header-attributes/card-header-attributes.component'
-import { AttributeComponent } from '../../../attribute/attribute.component'
+import { AttributeComponent } from '../../attribute/attribute.component'
 import { ChippedContentComponent } from '../../../chipped-content/chipped-content.component'
 import { componentTestSetup } from '@/test/helpers/component-test-setup'
 import { makeEducation } from '../../../data/__tests__/make-education'
