+++ src/app/resume-page/web-resume/experience-section/experience/experience.component.ts
@@ -1,6 +1,6 @@
 import { Component, computed, input } from '@angular/core'
 import { ChippedContentComponent } from '../../../chipped-content/chipped-content.component'
-import { AttributeComponent } from '../../../attribute/attribute.component'
+import { AttributeComponent } from '../../attribute/attribute.component'
 
 import { CardHeaderAttributesComponent } from '../../../card/card-header/card-header-attributes/card-header-attributes.component'
 import { DateRangeComponent } from '../../../date-range/date-range.component'
