+++ src/app/resume-page/web-resume/projects-section/project/project.component.spec.ts
@@ -13,7 +13,7 @@ import { byTestId } from '@/test/helpers/test-id'
 import { By } from '@angular/platform-browser'
 import { DateRange } from '../../../data/date-range'
 import { CardHeaderAttributesComponent } from '../../../card/card-header/card-header-attributes/card-header-attributes.component'
-import { AttributeComponent } from '../../../attribute/attribute.component'
+import { AttributeComponent } from '../../attribute/attribute.component'
 import { ChippedContentComponent } from '../../../chipped-content/chipped-content.component'
 import { componentTestSetup } from '@/test/helpers/component-test-setup'
 import { makeProject } from '../../../data/__tests__/make-project'
