+++ src/app/resume-page/web-resume/projects-section/project/project.component.ts
@@ -3,7 +3,7 @@ import { Project, Stack } from '../../../data/project'
 import { Apps, Dns, FullStackedBarChart } from '@/data/material-symbols'
 import { ChippedContent } from '../../../chipped-content/chipped-content'
 import { ChippedContentComponent } from '../../../chipped-content/chipped-content.component'
-import { AttributeComponent } from '../../../attribute/attribute.component'
+import { AttributeComponent } from '../../attribute/attribute.component'
 import { CardHeaderAttributesComponent } from '../../../card/card-header/card-header-attributes/card-header-attributes.component'
 import { DateRangeComponent } from '../../../date-range/date-range.component'
 import { CardHeaderDetailComponent } from '../../../card/card-header/card-header-detail/card-header-detail.component'
